integrations/deriv_connector.py
# integrations/deriv_connector.py

import os
import json
import threading
from datetime import datetime
from typing import Dict, Any, Optional
import websocket
import time
import logging

logger = logging.getLogger("DerivConnector")
logging.basicConfig(level=logging.INFO)

class DerivConnector:
    def __init__(self, api_token: str, app_id: int = 1089):
        self.api_token = api_token
        self.app_id = app_id
        self.ws_url = f"wss://ws.derivws.com/websockets/v3?app_id={app_id}"
        self.trade_log = []

    def place_trade_over_r_100(self, amount=1.0, duration=1):
        ws = websocket.WebSocket()
        ws.connect(self.ws_url)

        # Step 1: Authorize
        ws.send(json.dumps({"authorize": self.api_token}))
        time.sleep(1)  # Wait for authorization

        # Step 2: Send trade
        trade_payload = {
            "buy": 1,
            "price": amount,
            "parameters": {
                "amount": amount,
                "basis": "stake",
                "contract_type": "CALL",  # "CALL" = Over trade
                "currency": "USD",
                "duration": duration,
                "duration_unit": "t",
                "symbol": "R_100",
                "barrier": "0"
            }
        }

        ws.send(json.dumps(trade_payload))
        logger.info("✅ Trade sent!")

        # Optional: Print response
        response = ws.recv()
        logger.info(f"📩 Response: {response}")

        ws.close()
        self._log_trade(trade_payload, response)
        return {"response": response}

    def place_trade_under_r_100(self, amount=1.0, duration=1):
        ws = websocket.WebSocket()
        ws.connect(self.ws_url)

        # Step 1: Authorize
        ws.send(json.dumps({"authorize": self.api_token}))
        time.sleep(1)  # Wait for authorization

        # Step 2: Send trade
        trade_payload = {
            "buy": 1,
            "price": amount,
            "parameters": {
                "amount": amount,
                "basis": "stake",
                "contract_type": "PUT",  # "PUT" = Under trade
                "currency": "USD",
                "duration": duration,
                "duration_unit": "t",
                "symbol": "R_100",
                "barrier": "0"
            }
        }

        ws.send(json.dumps(trade_payload))
        logger.info("✅ Trade sent!")

        # Optional: Print response
        response = ws.recv()
        logger.info(f"📩 Response: {response}")

        ws.close()
        self._log_trade(trade_payload, response)
        return {"response": response}

    def _log_trade(self, trade_payload: Dict[str, Any], response: str):
        timestamp = datetime.utcnow().isoformat()
        entry = {
            "timestamp": timestamp,
            "trade": trade_payload,
            "response": response
        }
        self.trade_log.append(entry)
        os.makedirs("logs", exist_ok=True)
        with open("logs/trade_log.csv", "a") as log_file:
            log_file.write(f"{timestamp},{json.dumps(trade_payload)},{response}\n")

    # Add more methods as needed (account info, asset index, etc.)

# Sample usage (for testing)
if __name__ == "__main__":
    TEST_TOKEN = os.getenv("DERIV_TOKEN", "your-test-token")
    deriv = DerivConnector(api_token=TEST_TOKEN)
    deriv.place_trade_over_r_100(amount=1.0, duration=1)

# Create a global instance for use in other modules
DERIV_TOKEN = os.getenv("DERIV_TOKEN", "your-real-token")
deriv_connector = DerivConnector(api_token=DERIV_TOKEN)
