requirements.txt
numpy
pandas
scikit-learn
tensorflow
torch
xgboost
lightgbm

transformers
nltk
openai

streamlit
pyttsx3
SpeechRecognition
pyaudio
twilio
whatsapp-web
telethon

cryptography
pyjwt
pyopenssl
pycryptodome

gTTS
pydub
playsound

qiskit

flask
fastapi
requests
websockets

python-socketio
aiohttp
websocket-client

gunicorn
uvicorn
watchdog
psutil
schedule

matplotlib
seaborn
plotly
loguru

pyyaml
ujson
xmltodict
python-json-logger

pytest
black
isort
mypy

tqdm
joblib
multiprocess
colorama
rich